requirements.txt
appdirs==1.4.4
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
cffi==1.14.0
cryptography==2.9.2
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.1
jwt==1.0.0
more-itertools==4.2.0
pluggy==0.6.0
psycopg2-binary==2.8.5
py==1.5.3
pycparser==2.20
PyJWT==1.7.1
pytest==3.6.2
pytz==2020.1
six==1.11.0
sqlparse==0.3.1
virtualenv==20.0.25
whitenoise==5.1.0
zipp==3.1.0
